+++ src/configs/typeorm.config.ts
@@ -11,9 +11,9 @@ import { PostComplaints } from 'src/posts/entities/postComplaints.entity';
 import { ProcessState } from 'src/posts/entities/processState.entity';
 import { PostsLikeRecord } from 'src/posts/entities/postsLikeRecord.entity';
 import { PostsViewRecord } from 'src/posts/entities/postsViewRecord.entity';
-import { Followings } from 'src/mypage/followings.entity';
+import { Followings } from 'src/mypage/entities/followings.entity';
 import { PostImage } from 'src/posts/entities/postImage.entity';
-import { PurchaseHistory } from 'src/mypage/purchaseHistory.entity';
+import { PurchaseHistory } from 'src/mypage/entities/purchaseHistory.entity';
 import { Location } from 'src/users/entities/location.entity';
 import { ChatRoom } from 'src/chats/chatRoom.entity';
 import { Chat } from 'src/chats/chat.entity';
