+++ src/reviews/review.service.ts
@@ -1,6 +1,6 @@
 import { BadRequestException, ForbiddenException, Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { PurchaseHistory } from 'src/mypage/purchaseHistory.entity';
+import { PurchaseHistory } from 'src/mypage/entities/purchaseHistory.entity';
 import { User } from 'src/users/entities/user.entity';
 import { EntityManager, getConnection, getRepository } from 'typeorm';
 import { ReviewDto } from './dto/review.dto';
